fix some deployment issues - Human implement CursorGPT's coding
