--- .gitmeta/.gitmeta
@@ -1,23 +1,0 @@
-commit QmYroBxUzmuZ6ph9tCuerSifrDfSt3MJBJD8eaLk72MTgv
-Branch: PR-pwywModule
-Signed-off by: danimesq.eth (0xDDfC2e10702d8A781727A34D83B3bb3CA94a3E91)
-Signature: aab0a069fd6e892dcec9582c6e058779404a027453e71d7dd2a4f4bea89988b66ddde7a603debf0899626c089fe20d5ade6e219fb0cde073a7e6b3c2dd7e2db001
-
-------------------------------
-
-Commit CID if including git's SHA1: QmcrxLWsX9r51m69k1xELCDLGyjZxYkKkhjSaC3KWRWY1w
-Commit CID if containing only file hashes: QmQVdYVgamE1vD7cdPnFUk1yM32kZfyLWSxuvofbKE3X1C
-
-------------------------------
-
-.gitmeta version: r8
-Local folder: /home/daniella/Desktop/hey
-Remote Name: origin
-Remote URL: https://github.com/danimesq/hey.git
-
-------------------------------
-
-chown -h 1000:1000 'AIConversation.md'
-chmod 0664 'AIConversation.md'
-/usr/bin/touch -hcmd "2024-08-03 16:26:25.6892574220 -0300" 'AIConversation.md'
-/usr/bin/touch -hcad "2024-08-03 16:26:25.7572600380 -0300" 'AIConversation.md'
