--- .gitmeta/cids_detailed/with-sha1.txt
@@ -1,3 +1,0 @@
-QmbFMke1KXqnYyBBWxB74N4c5SBnJMVAiMNRcGu6x1AwQH
-52caca9c798374e81d0e91559f4477c3b0b9c97f
-QmPQm5HosDfKAMcoTVRhqBxxC4fj4T4QgwUiJmSSNmKEJB
